refactor: add PermissionToggle, remove duplicate code between role types
