+++ src/components/Groups/GroupRoles/GroupPermissionToggle.tsx
@@ -1,13 +1,8 @@
-import { Box, Switch, Typography } from "@mui/material";
-import { t } from "i18next";
-import { ChangeEvent } from "react";
 import {
   GroupRolePermissionInput,
   GroupRolePermissionsFragment,
 } from "../../../apollo/gen";
-import theme from "../../../styles/theme";
-import { getPermissionText } from "../../../utils/role.utils";
-import Flex from "../../Shared/Flex";
+import PermissionToggle from "../../Roles/PermissionToggle";
 
 interface Props {
   formValues: GroupRolePermissionInput;
@@ -22,47 +17,16 @@ const GroupPermissionToggle = ({
   permissions,
   setFieldValue,
 }: Props) => {
-  const { displayName, description, inDev } = getPermissionText(permissionName);
-  if (inDev) {
-    return null;
-  }
-
   const permissionInput = formValues[permissionName];
   const isEnabled = permissions[permissionName];
-  const isChecked = !!(permissionInput !== undefined
-    ? permissionInput
-    : isEnabled);
-
-  const handleSwitchChange = ({
-    target: { checked },
-  }: ChangeEvent<HTMLInputElement>) => {
-    if (!checked && isEnabled) {
-      setFieldValue(permissionName, false);
-      return;
-    }
-    if (checked === isEnabled) {
-      setFieldValue(permissionName, undefined);
-      return;
-    }
-    setFieldValue(permissionName, true);
-  };
 
   return (
-    <Flex justifyContent="space-between" marginBottom={2.8}>
-      <Box>
-        <Typography>{displayName}</Typography>
-
-        <Typography fontSize={12} sx={{ color: theme.palette.text.secondary }}>
-          {description}
-        </Typography>
-      </Box>
-
-      <Switch
-        inputProps={{ "aria-label": displayName || t("labels.switch") }}
-        onChange={handleSwitchChange}
-        checked={isChecked}
-      />
-    </Flex>
+    <PermissionToggle
+      permissionName={permissionName}
+      permissionInput={permissionInput}
+      setFieldValue={setFieldValue}
+      isEnabled={isEnabled}
+    />
   );
 };
